nixos/gitlab-runner: add dockerDevices option to executors
This option allows to hand host devices through to the docker containers
CI jobs are excuted in.
One example where this is useful is exposing USB devices to jobs, such
as an Android phone to a CI job that uses ADB to run an app on the
device.

See https://docs.gitlab.com/runner/executors/docker/#using-host-devices
for more information.